Refactor: Genericize AST node types
