+++ graphtage/__init__.py
@@ -6,7 +6,7 @@ from .edits import *
 
 from .version import __version__, VERSION_STRING
 from . import (
-    bounds, builder, constraints, dataclasses, edits, expressions, fibonacci, formatter, levenshtein, matching,
+    ast, bounds, builder, constraints, dataclasses, edits, expressions, fibonacci, formatter, levenshtein, matching,
     object_set, pickle, printer, pydiff, search, sequences, tree, utils
 )
 from . import csv, json, xml, yaml, plist
